fix typecheck
