+++ src/flows/shared/steps/OrganizeCollection/Editor/DragAndDrop/DroppableSection.tsx
@@ -33,7 +33,8 @@ export default function DroppableSection({ children, columns, id, items, style, 
     animateLayoutChanges,
   });
 
-  const { setActiveSectionIdState, addSection } = useCollectionEditorActions();
+  const { setActiveSectionIdState, addSection, deleteSection, setTokensIsSelected, unstageTokens } =
+    useCollectionEditorActions();
 
   // Set section as active on mousedown instead of on click so that starting to drag an item immediately activates that section
   const handleMouseDown = useCallback(() => {
@@ -46,6 +47,12 @@ export default function DroppableSection({ children, columns, id, items, style, 
   const handleAddSectionClick = useCallback(() => {
     addSection();
   }, [addSection]);
+
+  const handleDeleteSectionClick = useCallback(() => {
+    setTokensIsSelected(itemIds, false);
+    unstageTokens(itemIds);
+    deleteSection(id);
+  }, [deleteSection, id, itemIds, setTokensIsSelected, unstageTokens]);
 
   return (
     <>
@@ -63,12 +70,11 @@ export default function DroppableSection({ children, columns, id, items, style, 
             ...attributes,
             ...listeners,
           }}
-          id={id}
           columns={columns}
           {...props}
           isActive={isActive}
           isEmpty={items.length === 0}
-          itemIds={itemIds}
+          handleDeleteSectionClick={handleDeleteSectionClick}
         >
           {children}
         </Section>
